fix: Resolve all SwiftLint error-level violations for CI
- Split large test files (>300 lines) into smaller, focused modules:
  * CoreModuleCoverageTests (303→142 lines) → CoreModuleUtilityTests
  * LLMRateLimiterTests (319→163 lines) → LLMRateLimiterConfigTests
  * AllowanceCommonUseCaseTests (307→154 lines) → AllowanceRealWorldTests
  * AllowanceEdgeCaseTests (331→215 lines) → AllowancePersistenceEdgeCaseTests
  * ImageManagerTests (506 lines) → Split into 5 focused test files
  * DataServiceTest, ThemeManagerTests, QuizGenerationServiceTests, etc.

- Fixed isEmpty violations: Replaced count == 0 and count >= 0 with isEmpty checks
  * BasicEncryptionTests, CoreModuleCoverageTests, InsightsCoordinatorTest, CoreCoverageTests

- Fixed line length violations in BasicPDFGenerator.swift

- Extracted helper classes and mock objects into separate files for better organization

All error-level violations resolved. CI should now pass.
