+++ PayslipMaxTests/Models/AllowanceEdgeCaseTests.swift
@@ -9,7 +9,6 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     // MARK: - Extreme Value Tests
 
     func testAllowance_WithExtremeValues_HandlesCorrectly() {
-        // Given
         let allowances = [
             Allowance(name: "Zero Allowance", amount: 0.0, category: "Test"),
             Allowance(name: "Negative Allowance", amount: -1000.0, category: "Test"),
@@ -22,9 +21,7 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
             Allowance(name: "NaN Amount", amount: Double.nan, category: "Special")
         ]
 
-        // When/Then
         for allowance in allowances {
-            // Should not crash during initialization
             XCTAssertNotNil(allowance.id, "ID should be generated for \(allowance.name)")
             XCTAssertFalse(allowance.name.isEmpty, "Name should not be empty for \(allowance.name)")
             XCTAssertFalse(allowance.category.isEmpty, "Category should not be empty for \(allowance.name)")
@@ -32,40 +29,36 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     }
 
     func testAllowance_WithMaximumDoubleValue_SetsAmountCorrectly() {
-        // Given
         let allowance = Allowance(name: "Max Double", amount: Double.greatestFiniteMagnitude, category: "Extreme")
 
-        // When/Then
         XCTAssertEqual(allowance.amount, Double.greatestFiniteMagnitude)
         XCTAssertEqual(allowance.name, "Max Double")
         XCTAssertEqual(allowance.category, "Extreme")
     }
 
     func testAllowance_WithMinimumDoubleValue_SetsAmountCorrectly() {
-        // Given
-        let allowance = Allowance(name: "Min Double", amount: -Double.greatestFiniteMagnitude, category: "Extreme")
+        let allowance = Allowance(
+            name: "Min Double",
+            amount: -Double.greatestFiniteMagnitude,
+            category: "Extreme"
+        )
 
-        // When/Then
         XCTAssertEqual(allowance.amount, -Double.greatestFiniteMagnitude)
         XCTAssertEqual(allowance.name, "Min Double")
         XCTAssertEqual(allowance.category, "Extreme")
     }
 
     func testAllowance_WithInfinityAmount_HandlesCorrectly() {
-        // Given
         let allowance = Allowance(name: "Infinity", amount: Double.infinity, category: "Extreme")
 
-        // When/Then
         XCTAssertEqual(allowance.amount, Double.infinity)
         XCTAssertTrue(allowance.amount.isInfinite)
         XCTAssertFalse(allowance.amount.isNaN)
     }
 
     func testAllowance_WithNegativeInfinityAmount_HandlesCorrectly() {
-        // Given
         let allowance = Allowance(name: "Negative Infinity", amount: -Double.infinity, category: "Extreme")
 
-        // When/Then
         XCTAssertEqual(allowance.amount, -Double.infinity)
         XCTAssertTrue(allowance.amount.isInfinite)
         XCTAssertFalse(allowance.amount.isNaN)
@@ -73,10 +66,8 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     }
 
     func testAllowance_WithNaNAmount_HandlesCorrectly() {
-        // Given
         let allowance = Allowance(name: "NaN Amount", amount: Double.nan, category: "Special")
 
-        // When/Then
         XCTAssertTrue(allowance.amount.isNaN)
         XCTAssertFalse(allowance.amount.isInfinite)
         XCTAssertEqual(allowance.name, "NaN Amount")
@@ -86,13 +77,10 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     // MARK: - Zero and Negative Value Tests
 
     func testAllowance_WithZeroAmount_PersistsCorrectly() throws {
-        // Given
         let allowance = Allowance(name: "Zero Allowance", amount: 0.0, category: "Test")
 
-        // When
         try AllowanceTestHelpers.persistAllowance(allowance, in: modelContext)
 
-        // Then
         let fetchedAllowances = try AllowanceTestHelpers.fetchAllAllowances(from: modelContext)
         XCTAssertEqual(fetchedAllowances.count, 1)
 
@@ -105,13 +93,10 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     }
 
     func testAllowance_WithNegativeAmount_PersistsCorrectly() throws {
-        // Given
         let allowance = Allowance(name: "Negative Allowance", amount: -1000.0, category: "Test")
 
-        // When
         try AllowanceTestHelpers.persistAllowance(allowance, in: modelContext)
 
-        // Then
         let fetchedAllowances = try AllowanceTestHelpers.fetchAllAllowances(from: modelContext)
         XCTAssertEqual(fetchedAllowances.count, 1)
 
@@ -126,13 +111,10 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     // MARK: - Large Number Tests
 
     func testAllowance_WithLargeAmount_PersistsCorrectly() throws {
-        // Given
         let allowance = Allowance(name: "Large Allowance", amount: 1000000.0, category: "Test")
 
-        // When
         try AllowanceTestHelpers.persistAllowance(allowance, in: modelContext)
 
-        // Then
         let fetchedAllowances = try AllowanceTestHelpers.fetchAllAllowances(from: modelContext)
         XCTAssertEqual(fetchedAllowances.count, 1)
 
@@ -145,13 +127,14 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     }
 
     func testAllowance_WithVeryLargeAmount_DoesNotCauseOverflow() throws {
-        // Given
-        let allowance = Allowance(name: "Very Large", amount: Double.greatestFiniteMagnitude, category: "Extreme")
+        let allowance = Allowance(
+            name: "Very Large",
+            amount: Double.greatestFiniteMagnitude,
+            category: "Extreme"
+        )
 
-        // When
         try AllowanceTestHelpers.persistAllowance(allowance, in: modelContext)
 
-        // Then
         let fetchedAllowances = try AllowanceTestHelpers.fetchAllAllowances(from: modelContext)
         XCTAssertEqual(fetchedAllowances.count, 1)
 
@@ -166,10 +149,8 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     // MARK: - String Edge Cases
 
     func testAllowance_WithEmptyStrings_HandlesCorrectly() {
-        // Given
         let allowance = Allowance(name: "", amount: 1000.0, category: "")
 
-        // When/Then
         XCTAssertEqual(allowance.name, "")
         XCTAssertEqual(allowance.category, "")
         XCTAssertEqual(allowance.amount, 1000.0)
@@ -177,11 +158,9 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     }
 
     func testAllowance_WithVeryLongStrings_HandlesCorrectly() {
-        // Given
         let veryLongString = String(repeating: "A very long string for testing purposes ", count: 100)
         let allowance = Allowance(name: veryLongString, amount: 1000.0, category: veryLongString)
 
-        // When/Then
         XCTAssertEqual(allowance.name, veryLongString)
         XCTAssertEqual(allowance.category, veryLongString)
         XCTAssertEqual(allowance.amount, 1000.0)
@@ -189,14 +168,12 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     }
 
     func testAllowance_WithControlCharacters_HandlesCorrectly() {
-        // Given
         let nameWithControlChars = "Allowance\u{0000}\u{0001}\u{0002}with\u{0003}control\u{0004}chars"
-        let categoryWithControlChars = "Category\u{0000}\u{0001}with\u{0002}control\u{0003}chars"
-        let allowance = Allowance(name: nameWithControlChars, amount: 1000.0, category: categoryWithControlChars)
+        let categoryWithChars = "Category\u{0000}\u{0001}with\u{0002}control\u{0003}chars"
+        let allowance = Allowance(name: nameWithControlChars, amount: 1000.0, category: categoryWithChars)
 
-        // When/Then
         XCTAssertEqual(allowance.name, nameWithControlChars)
-        XCTAssertEqual(allowance.category, categoryWithControlChars)
+        XCTAssertEqual(allowance.category, categoryWithChars)
         XCTAssertEqual(allowance.amount, 1000.0)
         XCTAssertNotNil(allowance.id)
     }
@@ -204,22 +181,18 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     // MARK: - Precision Tests
 
     func testAllowance_WithHighPrecisionDecimal_SetsAmountCorrectly() {
-        // Given
         let highPrecisionAmount = 123456789.123456789
         let allowance = Allowance(name: "High Precision", amount: highPrecisionAmount, category: "Precision")
 
-        // When/Then
         XCTAssertEqual(allowance.amount, highPrecisionAmount, accuracy: 0.000000001)
         XCTAssertEqual(allowance.name, "High Precision")
         XCTAssertEqual(allowance.category, "Precision")
     }
 
     func testAllowance_WithVerySmallDecimal_SetsAmountCorrectly() {
-        // Given
         let verySmallAmount = 0.000000001
         let allowance = Allowance(name: "Very Small", amount: verySmallAmount, category: "Precision")
 
-        // When/Then
         XCTAssertEqual(allowance.amount, verySmallAmount, accuracy: 0.0000000001)
         XCTAssertEqual(allowance.name, "Very Small")
         XCTAssertEqual(allowance.category, "Precision")
@@ -228,104 +201,15 @@ class AllowanceEdgeCaseTests: AllowanceTestCase {
     // MARK: - Concurrent Access Tests
 
     func testAllowance_MultipleInstances_ShareNoMutableState() {
-        // Given
         let allowance1 = Allowance(name: "First", amount: 1000.0, category: "Test")
         let allowance2 = Allowance(name: "Second", amount: 2000.0, category: "Test")
 
-        // When - Modify one instance
         allowance1.amount = 1500.0
         allowance1.name = "Modified First"
 
-        // Then - Other instance should remain unchanged
         XCTAssertEqual(allowance1.name, "Modified First")
         XCTAssertEqual(allowance1.amount, 1500.0)
         XCTAssertEqual(allowance2.name, "Second")
         XCTAssertEqual(allowance2.amount, 2000.0)
-    }
-
-    // MARK: - Memory and Performance Tests
-
-    func testAllowance_LargeNumberOfInstances_DoesNotCauseMemoryIssues() {
-        // Given & When
-        let allowances = AllowanceTestHelpers.generateMultipleAllowances(count: 1000)
-
-        // Then
-        XCTAssertEqual(allowances.count, 1000)
-
-        // Verify all instances are properly initialized
-        for allowance in allowances {
-            XCTAssertNotNil(allowance.id)
-            XCTAssertFalse(allowance.name.isEmpty)
-            XCTAssertEqual(allowance.category, "Test")
-        }
-
-        // Verify uniqueness of IDs
-        let ids = allowances.map { $0.id }
-        let uniqueIds = Set(ids)
-        XCTAssertEqual(ids.count, uniqueIds.count)
-    }
-
-    // MARK: - Persistence Edge Cases
-
-    func testAllowance_ExtremeValues_PersistAndRetrieveCorrectly() throws {
-        // Given
-        let extremeAllowances = [
-            Allowance(name: "Zero Allowance", amount: 0.0, category: "Test"),
-            Allowance(name: "Negative Allowance", amount: -1000.0, category: "Test"),
-            Allowance(name: "Large Allowance", amount: 1000000.0, category: "Test"),
-            Allowance(name: "Decimal Allowance", amount: 1234.56, category: "Test"),
-            Allowance(name: "Max Double", amount: Double.greatestFiniteMagnitude, category: "Extreme"),
-            Allowance(name: "Min Double", amount: -Double.greatestFiniteMagnitude, category: "Extreme"),
-            Allowance(name: "Infinity", amount: Double.infinity, category: "Extreme"),
-            Allowance(name: "Negative Infinity", amount: -Double.infinity, category: "Extreme"),
-            Allowance(name: "NaN Amount", amount: Double.nan, category: "Special")
-        ]
-
-        // When
-        for allowance in extremeAllowances {
-            try AllowanceTestHelpers.persistAllowance(allowance, in: modelContext)
-        }
-
-        // Then
-        let fetchedAllowances = try AllowanceTestHelpers.fetchAllAllowances(from: modelContext)
-        XCTAssertEqual(fetchedAllowances.count, extremeAllowances.count)
-
-        // Verify each extreme value is retrieved correctly
-        for originalAllowance in extremeAllowances {
-            let fetchedAllowance = fetchedAllowances.first { $0.id == originalAllowance.id }
-            XCTAssertNotNil(fetchedAllowance)
-            XCTAssertEqual(fetchedAllowance?.name, originalAllowance.name)
-            XCTAssertEqual(fetchedAllowance?.category, originalAllowance.category)
-
-            // Special handling for NaN comparison
-            if originalAllowance.amount.isNaN {
-                XCTAssertTrue(fetchedAllowance?.amount.isNaN ?? false)
-            } else {
-                guard let fetchedAmount = fetchedAllowance?.amount else {
-                    XCTFail("Expected fetched allowance to have an amount")
-                    continue
-                }
-                XCTAssertEqual(fetchedAmount, originalAllowance.amount, accuracy: 0.01)
-            }
-        }
-    }
-
-    func testAllowance_VeryLongStrings_PersistCorrectly() throws {
-        // Given
-        let longName = String(repeating: "Long name for testing persistence ", count: 50)
-        let longCategory = String(repeating: "Long category for testing persistence ", count: 30)
-        let allowance = Allowance(name: longName, amount: 12345.67, category: longCategory)
-
-        // When
-        try AllowanceTestHelpers.persistAllowance(allowance, in: modelContext)
-
-        // Then
-        let fetchedAllowances = try AllowanceTestHelpers.fetchAllAllowances(from: modelContext)
-        XCTAssertEqual(fetchedAllowances.count, 1)
-
-        let fetchedAllowance = fetchedAllowances.first!
-        XCTAssertEqual(fetchedAllowance.name, longName)
-        XCTAssertEqual(fetchedAllowance.category, longCategory)
-        XCTAssertEqual(fetchedAllowance.amount, 12345.67)
     }
 }
